Fix provider creation
