+++ src/class/provider/index.js
@@ -1,2 +1,1 @@
-export { default as DebounceProvider } from './DebounceProvider';
 export { default as providerFactory } from './providerFactory';
